Rename PostalCodeLookup method to LookupPostalCode.
The fact that there is a data contract object named PostalCodeLookup with
the same name as a PostalCodeLookup method bothers me. Renamed the method
so that it better describes the behavior (verb) of the method, even
though it does not match the name of the actual service invoked on
geonames. Kept the arguments object name as a noun, since it has a data
contract.